test: PHP 8.3 unit suite curation for zero failures/errors
Align production and test code with PHP 8.x strictness (casts, Phrase
messages, cron/indexer edge cases) and restore ObjectManager harness
compatibility for legacy constructor argument names (auth/session/helper).

Update unit test fixtures and expectations where generator output or mock
behavior changed on PHP 8.3 / PHPUnit 12.

+++ setup/src/Magento/Setup/Fixtures/AttributeSet/SwatchesGenerator.php
@@ -113,7 +113,7 @@ class SwatchesGenerator
      */
     private function generateSwatchColor($index)
     {
-        return '#' . str_repeat(dechex(255 * $index), 3);
+        return '#' . str_repeat(dechex((int) (255 * $index)), 3);
     }
 
     /**
